tests/wpt/mozilla/tests/mozilla/bluetooth/requestDevice/canonicalizeFilter/empty-namePrefix.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

let test_specs = [{
  filters: [{ namePrefix: ''}]
}, {
  filters: [{ namePrefix: '', name: 'Name'}]
}, {
  filters: [{ namePrefix: '', services: ['heart_rate']}]
}, {
  filters: [{ namePrefix: '', name: 'Name', services: ['heart_rate']}]
}, {
  filters: [{ namePrefix: ''}],
  optionalServices: ['heart_rate']
}, {
  filters: [{ namePrefix: '', name: 'Name'}],
  optionalServices: ['heart_rate']
}, {
  filters: [{ namePrefix: '', services: ['heart_rate']}],
  optionalServices: ['heart_rate']
}, {
  filters: [{ namePrefix: '', name: 'Name', services: ['heart_rate']}],
  optionalServices: ['heart_rate']
}];

test(function() {
    window.testRunner.setBluetoothMockDataSet('EmptyAdapter');
    test_specs.forEach(args => {
        assert_throws(new TypeError(), function () {
            window.navigator.bluetooth.requestDevice(args);
        });
    });
}, 'requestDevice with empty namePrefix. Should reject with TypeError.');
</script>
